Changes again 2

+++ src/components/landing/admin-dashboard/HolidayPreviewModal.tsx
@@ -109,7 +109,7 @@ export const HolidayPreviewModal: React.FC<HolidayPreviewModalProps> = ({
                     <div className="text-left flex-1">
                       <p className="font-medium text-gray-800">{user.username}</p>
                       <p className="text-xs text-gray-600">
-                        Registered: {new Date(user.registration_date).toLocaleDateString('en-GB', { day: 'short', month: 'short', year: '2-digit' })}
+                        Registered: {new Date(user.registration_date).toLocaleDateString('en-GB', { day: 'numeric', month: 'short', year: '2-digit' })}
                       </p>
                     </div>
                     <div className="flex items-center gap-3">
@@ -131,13 +131,13 @@ export const HolidayPreviewModal: React.FC<HolidayPreviewModalProps> = ({
                         <div>
                           <span className="text-gray-600">Current last date</span>
                           <p className="font-semibold text-gray-800">
-                            {new Date(user.last_date).toLocaleDateString('en-GB', { day: 'short', month: 'short', year: '2-digit' })}
+                            {new Date(user.last_date).toLocaleDateString('en-GB', { day: 'numeric', month: 'short', year: '2-digit' })}
                           </p>
                         </div>
                         <div>
                           <span className="text-gray-600">New last date</span>
                           <p className="font-semibold text-green-700">
-                            {new Date(user.new_last_date).toLocaleDateString('en-GB', { day: 'short', month: 'short', year: '2-digit' })}
+                            {new Date(user.new_last_date).toLocaleDateString('en-GB', { day: 'numeric', month: 'short', year: '2-digit' })}
                           </p>
                         </div>
                       </div>
